Add test for Carrier status persistence
